requirements.txt
Flask
pymongo
chromadb
sentence-transformers
beautifulsoup4
Flask-Login
Flask-WTF
Flask-Limiter
Flask-Cors
pydantic
python-dotenv
